build(typescript): convert js to ts, change folder structure sutable for typescript
