--- src/App.css
@@ -1,24 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 30vmin;
-  pointer-events: none;
-  margin-bottom: 8vmin;
-}
-
-.App-header {
-  background-color: #fff8e1;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: #001743;
-}
-
-.App-link {
-  color: #206dfa;
-}
