+++ src/app/app.tsx
@@ -1,0 +1,11 @@
+import { observer } from "mobx-react";
+
+export const App = observer(() => {
+
+
+  return (
+    <div className="App">
+
+    </div>
+  );
+});
